requirements.txt
pyyaml
jproperties
mcdreforged